Add more tests and test utils
