+++ tests/test_utils.py
@@ -1,5 +1,7 @@
+import networkx as nx
 import numpy as np
 import skimage as sk
+from cell_tracking_metrics.tracking_graph import TrackingGraph
 
 
 def get_annotated_image(img_size=256, num_labels=3, sequential=True, seed=1):
@@ -74,3 +76,30 @@ def get_annotated_movie(
                 y[frame, :, :][label_loc] = new_label
 
     return y.astype("int32")
+
+
+def get_movie_with_graph(ndims=3, n_frames=3, n_labels=3):
+    movie = get_annotated_movie(
+        labels_per_frame=n_labels, frames=n_frames, mov_type="repeated"
+    )
+
+    # Extend to 3d if needed
+    if ndims == 4:
+        movie = np.stack([movie, movie, movie], axis=-1)
+
+    # We can assume each object is present and connected across each frame
+    G = nx.DiGraph()
+    for t in range(n_frames - 1):
+        for i in range(1, n_labels + 1):
+            G.add_edge(f"{i}_{t}", f"{i}_{t+1}")
+
+    attrs = {}
+    for t in range(n_frames):
+        for i in range(1, n_labels + 1):
+            a = {"t": t, "y": 0, "x": 0, "segmentation_id": i}
+            if ndims == 4:
+                a["z"] = 0
+            attrs[f"{i}_{t}"] = a
+    nx.set_node_attributes(G, attrs)
+
+    return TrackingGraph(G), movie
